requirements.txt
msgpack-python
greenlet
python-dateutil
colorama
click
click-plugins
click-shell
psutil
requests
grequests
PyYAML
six
Cerberus
simplejson
verboselogs
coloredlogs
configobj
versioneer
autopep8
